modules/bluetooth_tools.py
try:
    import bluetooth
except ImportError:
    bluetooth = None

import threading
import time

class BluetoothAttacker:
    def __init__(self):
        self.running = False

    def scan_devices(self):
        if not bluetooth:
            return ["Error: pybluez not installed (pip install pybluez)"]
        
        try:
            nearby_devices = bluetooth.discover_devices(lookup_names=True)
            return [f"{addr} - {name}" for addr, name in nearby_devices]
        except Exception as e:
            return [f"Scan Error: {e}"]

    def start_bluejacking(self, target_addr, message_file):
        if not bluetooth: return "Error: pybluez missing"
        
        # Simulating bluejacking (Sending contact card or message)
        # Real bluejacking is obsolete on modern phones but we can try sending a file/msg via OBEX
        # For this tool, we will try to connect and send a simple message string if service exists
        
        try:
            with open(message_file, 'r') as f:
                data = f.read()
        except:
            data = "You have been hacked by AntiLife"

        self.running = True
        threading.Thread(target=self._spam_msg, args=(target_addr, data)).start()
        return "Bluejacking initiated..."

    def _spam_msg(self, addr, date):
        # This is a placeholder for the actual OBEX push or L2CAP connection
        # Creating a real bluejacker requires complex protocol handling
        print(f"Sending to {addr}...")
        time.sleep(2)
        print("Sent.")
