requirements.txt
customtkinter
scapy
requests
flask
Pillow
